+++ Domain/DTOs/Phone/PhoneOutput.cs
@@ -13,7 +13,7 @@ namespace Domain.DTOs.Phone
         public int Id { get; set; }
         [Required(ErrorMessage = "El campo {0} es requerido")]
         [StringLength(30, MinimumLength = 1, ErrorMessage = "El Máximo es de 30 carateres")]
-        [Range(0, int.MaxValue, ErrorMessage = "Ingresa solo números")]
+        [Range(0, double.MaxValue, ErrorMessage = "Ingresa solo números")]
         [Display(Name = "Teléfono")]
         public string phone { get; set; }
